sync-scripts/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
Flask==2.3.3
python-dotenv==1.0.0
schedule==1.2.0
